Rename module that was causing import errors in rare cases
Only failed in snake_oil in snake_oil_simulator.py script in the
run_ert2_test_data_setups.yml. It happens when it imports from the res
package: from res.test.synthesizer import OilSimulator. Unclear why
this would occur, but it is necessary to do before changing the
behaviour of ErtAdapter in the following commit.
