+++ ert3/engine/_run.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Tuple
 import ert
 import ert3
 from ert3.config import SourceNS
-from ert_shared.asyncio import get_event_loop
+from ert_shared.async_utils import get_event_loop
 from ert_shared.ensemble_evaluator.ensemble.builder import create_step_builder
 
 from ._entity import TransmitterCoroutine
