+++ ert_shared/ensemble_evaluator/entity/function_step.py
@@ -3,7 +3,7 @@ import pickle
 from typing import Dict
 import prefect
 
-from ert_shared.asyncio import get_event_loop
+from ert_shared.async_utils import get_event_loop
 from ert_shared.ensemble_evaluator.client import Client
 from ert_shared.ensemble_evaluator.entity import identifiers as ids
