+++ ert_shared/ensemble_evaluator/entity/unix_step.py
@@ -7,7 +7,7 @@ import os
 
 import prefect
 
-from ert_shared.asyncio import get_event_loop
+from ert_shared.async_utils import get_event_loop
 from ert_shared.ensemble_evaluator.client import Client
 from ert_shared.ensemble_evaluator.entity import identifiers as ids
