+++ tests/ert_tests/ensemble_evaluator/prefect/conftest.py
@@ -14,7 +14,7 @@ import ert
 import ert_shared.ensemble_evaluator.ensemble.builder as ee
 from ensemble_evaluator_utils import _mock_ws
 
-from ert_shared.asyncio import get_event_loop
+from ert_shared.async_utils import get_event_loop
 from ert_shared.ensemble_evaluator.client import Client
 from ert_shared.ensemble_evaluator.config import EvaluatorServerConfig
