+++ tests/ert_tests/ensemble_evaluator/prefect/test_prefect_step.py
@@ -11,7 +11,7 @@ from functools import partial
 import ert
 from ert_utils import tmp
 import ert_shared.ensemble_evaluator.ensemble.builder as ee
-from ert_shared.asyncio import get_event_loop
+from ert_shared.async_utils import get_event_loop
 from ert_shared.ensemble_evaluator.entity import identifiers as ids
 from ert_shared.ensemble_evaluator.ensemble.prefect import PrefectEnsemble
