+++ tests/ert_tests/ert3/algorithms/test_fast.py
@@ -9,7 +9,7 @@ import ert3
 # Expected values for S1 and ST for ishigami function
 # sample_size = 15000, harmonics = 4, bounds = [-pi, pi]
 # uniform distribution
-from ert_shared.asyncio import get_event_loop
+from ert_shared.async_utils import get_event_loop
 
 ISHIGAMI_S1 = [0.3076, 0.4424, 6.351e-27]
 ISHIGAMI_ST = [0.5507, 0.4695, 0.2392]
