+++ tests/ert_tests/ert3/evaluator/test_evaluator.py
@@ -5,7 +5,7 @@ import pytest
 import ert
 import ert3
 from ert_shared.ensemble_evaluator.ensemble.builder import create_step_builder
-from ert_shared.asyncio import get_event_loop
+from ert_shared.async_utils import get_event_loop
 
 TEST_PARAMETRIZATION = [
     ([(0, 0, 0)], [[0] * 10]),
